#127: Removed unnecessary 2nd language test class and added a new test case to the language test class.
